restruct(fm-analysis)

+++ internal/tenant/fm/deals_transactions_types.go
@@ -6,10 +6,10 @@ const (
 )
 
 type DealTransactionsSummary struct {
-	TotalAmount   int64 `json:"total_amount"`   // итоговая сумма (доходы - расходы)
-	IncomeAmount  int64 `json:"income_amount"`  // сумма доходных транзакций
-	ExpenseAmount int64 `json:"expense_amount"` // сумма расходных транзакций
-	IncomeCount   int64 `json:"income_count"`   // количество доходных транзакций
-	ExpenseCount  int64 `json:"expense_count"`  // количество расходных транзакций
-	TotalCount    int64 `json:"total_count"`    // общее количество транзакций
+	TotalAmount   float64 `json:"total_amount"`   // итоговая сумма (доходы - расходы)
+	IncomeAmount  float64 `json:"income_amount"`  // сумма доходных транзакций
+	ExpenseAmount float64 `json:"expense_amount"` // сумма расходных транзакций
+	IncomeCount   int64   `json:"income_count"`   // количество доходных транзакций
+	ExpenseCount  int64   `json:"expense_count"`  // количество расходных транзакций
+	TotalCount    int64   `json:"total_count"`    // общее количество транзакций
 }
